+++ typings/local.d.ts
@@ -2,6 +2,6 @@
 /// <reference path="local/require/require.d.ts" />
 /// <reference path="local-modules/intro.js/index.d.ts" />
 /// <reference path="local-modules/jointjs/index.d.ts" />
-/// <reference path="local-modules/springy/index.d.ts" />
 /// <reference path="local-modules/n3/index.d.ts" />
 /// <reference path="local-modules/handlebars/index.d.ts" />
+/// <reference path="local-modules/webcola/index.d.ts" />